* master: (87 commits)
  src/common.h : Increase SF_MAX_CHANNELS from 256 to 1024.
  src/ : Add more checks for suported channel counts.
  AUTHORS : Add contributors.
  configure.ac: Fix detection of MinGW
  Purge redundant file binheader_readf_check.py.
  CMakeLists.txt : Use bash to run test_wrapper.sh.
  More fixes for HAVE_SNDIO_H changes.
  configure.ac : Fix chmod of Scripts/build-test-tarball.mk.
  CMakeLists.txt : Generate tests/test_wrapper.sh during configure.
  CMake : Build system nears parity with autotool build system.
  Update .gitignore for CMake artifacts.
  tests/ : Fix warnings from clang 3.5.
  Fix detection and usage OpenBSD's <sndio.h>.
  src/sndfile.hh: initialize all member variables
  CMake : Detect functions and libraries.
  CMakeLists.txt : Build a versioned shared library.
  CMake : Add all the library C files and start on CFLAGS.
  CMake : Split functions out into a separate file.
  CMake : Use check_type_size as needed.
  CMake : Detect and set COMPILER_IS_GCC.
  ...

+++ src/sd2.c
@@ -391,7 +391,7 @@ read_rsrc_int (const SD2_RSRC *prsrc, int offset)
 {	const unsigned char * data = prsrc->rsrc_data ;
 	if (offset < 0 || offset + 3 >= prsrc->rsrc_len)
 		return 0 ;
-	return (data [offset] << 24) + (data [offset + 1] << 16) + (data [offset + 2] << 8) + data [offset + 3] ;
+	return (((uint32_t) data [offset]) << 24) + (data [offset + 1] << 16) + (data [offset + 2] << 8) + data [offset + 3] ;
 } /* read_rsrc_int */
 
 static inline int
@@ -402,9 +402,9 @@ read_rsrc_marker (const SD2_RSRC *prsrc, int offset)
 		return 0 ;
 
 	if (CPU_IS_BIG_ENDIAN)
-		return (data [offset] << 24) + (data [offset + 1] << 16) + (data [offset + 2] << 8) + data [offset + 3] ;
+		return (((uint32_t) data [offset]) << 24) + (data [offset + 1] << 16) + (data [offset + 2] << 8) + data [offset + 3] ;
 	if (CPU_IS_LITTLE_ENDIAN)
-		return data [offset] + (data [offset + 1] << 8) + (data [offset + 2] << 16) + (data [offset + 3] << 24) ;
+		return data [offset] + (data [offset + 1] << 8) + (data [offset + 2] << 16) + (((uint32_t) data [offset + 3]) << 24) ;
 
 	return 0 ;
 } /* read_rsrc_marker */
